Move exceptions test to its own file
